Remove unused managers directory and apply clippy fixes
- Delete entire managers directory with unused wrappers
- Remove managers module declaration from lib.rs
- Apply automatic clippy fixes for better code quality

+++ src/utils/integrity/consistency_checker.rs
@@ -255,7 +255,7 @@ impl ConsistencyChecker {
                         "Page {} is allocated but not reachable from root",
                         page_id
                     ),
-                    affected_pages: vec![page_id as u64],
+                    affected_pages: vec![page_id],
                     severity: ErrorSeverity::Error,
                 });
             }
